refactor: introduce RootView
- removed props from GestureHandlerRootView as it is not intended
  to be handled as root view for renderings
- created a new RootView component to handle the initial font loading
  and splash screen hiding
- integrated RootView into the main application structure after provider
  and contexts before first rendering, to be in place where first needed
  - this way it could be a positive side effect, that the splash screen is
    visible for a longer time, as things will be setup first now
